+++ src/test/converter2/behavior/linkResolutionTs.ts
@@ -1,0 +1,6 @@
+export * from "./linkResolution";
+
+import { A as AnotherName } from "./linkResolution";
+
+/** {@link AnotherName | A!}{@link AnotherName A2!}{@link AnotherName} */
+export const localSymbolRef = 1;